nearly going

+++ ddl-scripts/create-history-tables.sql
@@ -1,5 +1,5 @@
 
 CREATE TABLE IF NOT EXISTS public.btc_history (
       datetime timestamp NOT NULL,
-      amount double NOT NULL,
+      amount double precision NOT NULL,
       PRIMARY KEY (datetime));
